# platform/internal/apigw/assets/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "USSLP Platform API",
    "version": "1.0.0",
    "summary": "The public front door of the Universal Smart Shelf Label Platform.",
    "description": "Every request is bound to exactly one tenant, and the tenant is derived from the credential rather than from anything in the request. An identifier belonging to another tenant is reported as `404 not_found`, never `403`: confirming that an identifier exists is itself a cross-tenant disclosure, so the API declines to distinguish 'you may not see this' from 'this does not exist'.\n\nErrors share one body: `{\"code\", \"message\", \"request_id\"}`. Quote the `request_id` in a support ticket and it joins the access log, the distributed trace and the audit record.\n\nRate limits are token buckets. `X-RateLimit-Limit` is the bucket depth, `X-RateLimit-Remaining` the tokens left, `X-RateLimit-Reset` the seconds until the bucket is full again, and `X-RateLimit-Bucket` says which of the three tiers (tenant, credential, expensive) is binding. A `429` carries `Retry-After` in seconds.\n\nRetries are safe on GET, HEAD, OPTIONS, PUT and DELETE. POST is not retried by the gateway; supply an `idempotency_key` where the schema offers one and a client-side retry becomes a no-op all the way down to the event store.",
    "contact": { "name": "USSLP Platform Team" },
    "license": { "name": "Proprietary", "identifier": "LicenseRef-USSLP-Commercial" }
  },
  "servers": [
...
  ],
...
  "security": [
    { "apiKey": [] },
    { "bearerAuth": [] },
    { "mutualTLS": [] }
  ],
  "paths": {
...
    "/v1/ota/jobs": {
      "get": {
        "operationId": "listOTAJobs",
        "summary": "Firmware rollout jobs.",
        "tags": ["ota"],
        "parameters": [
          {
            "name": "store_id", "in": "query", "required": false,
            "description": "Restrict to rollouts touching one store.",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "$ref": "#/components/responses/Document" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      },
      "post": {
        "operationId": "createOTAJob",
        "summary": "Start a staged firmware rollout.",
        "description": "Rollouts advance through cohorts and honour quiet hours; creating one requires `ota:admin` because its blast radius is the tenant's whole fleet.",
        "tags": ["ota"],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/Document" },
          "200": { "$ref": "#/components/responses/Document" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/ota/jobs/{jobId}": {
      "get": {
        "operationId": "getOTAJob",
        "summary": "One rollout's progress.",
        "tags": ["ota"],
        "parameters": [{ "$ref": "#/components/parameters/jobId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Document" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/ota/jobs/{jobId}/devices": {
      "get": {
        "operationId": "listOTAJobDevices",
        "summary": "Per-device outcomes within a rollout.",
        "tags": ["ota"],
        "parameters": [{ "$ref": "#/components/parameters/jobId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Document" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/ota/jobs/{jobId}/pause": {
      "post": {
        "operationId": "pauseOTAJob",
        "summary": "Pause a rollout.",
        "tags": ["ota"],
        "parameters": [{ "$ref": "#/components/parameters/jobId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Document" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/ota/jobs/{jobId}/resume": {
      "post": {
        "operationId": "resumeOTAJob",
        "summary": "Resume a paused rollout.",
        "tags": ["ota"],
        "parameters": [{ "$ref": "#/components/parameters/jobId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Document" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/ota/jobs/{jobId}/abort": {
      "post": {
        "operationId": "abortOTAJob",
        "summary": "Abort a rollout.",
        "tags": ["ota"],
        "parameters": [{ "$ref": "#/components/parameters/jobId" }],
        "responses": {
          "200": { "$ref": "#/components/responses/Document" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
    "/v1/ota/jobs/{jobId}/rollback": {
      "post": {
        "operationId": "rollbackOTAJob",
        "summary": "Roll a fleet back to the previous firmware.",
        "tags": ["ota"],
        "parameters": [{ "$ref": "#/components/parameters/jobId" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ReasonRequest" } } }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/Document" },
          "default": { "$ref": "#/components/responses/Error" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "A tenant-scoped API key presented as `Authorization: Bearer usslp_live_…`. The prefix is deliberate: a leaked key is findable by a secret scanner and by `git grep`."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "An ES256 token from the platform's token service. Keys are resolved by `kid` from a published JWKS, which is what makes rotation a key-set change rather than a redeployment."
      },
      "mutualTLS": {
        "type": "mutualTLS",
        "description": "A tenant client certificate from the USSLP hierarchy. The production path for machine-to-machine traffic: nothing bearer-shaped crosses the wire, so nothing bearer-shaped can be replayed from a log."
      }
    },
    "parameters": {
...
      "jobId": {
        "name": "jobId", "in": "path", "required": true,
        "description": "A firmware rollout job identifier.",
        "schema": { "type": "string", "maxLength": 128 }
      },
...
    },
    "responses": {
      "Document": {
        "description": "The upstream service's document, relayed unchanged.",
        "content": { "application/json": { "schema": { "type": "object" } } }
      },
      "Error": {
        "description": "An error. `400` invalid argument; `401` no or bad credential; `403` the credential lacks the permission; `404` no such resource, including one belonging to another tenant or outside the credential's store scope; `405` wrong method; `413` body too large; `429` rate limited, with `Retry-After`; `502`/`503`/`504` an upstream was unreachable, circuit-broken or slow.",
        "headers": {
          "X-Request-Id": {
            "description": "Correlates this response with the access log, the trace and the audit record.",
            "schema": { "type": "string" }
          },
          "Retry-After": {
            "description": "Seconds to wait before retrying. Present on 429 and on a circuit-broken 503.",
            "schema": { "type": "integer", "minimum": 1 }
          },
          "X-RateLimit-Limit": { "description": "Token bucket depth.", "schema": { "type": "integer" } },
          "X-RateLimit-Remaining": { "description": "Tokens left.", "schema": { "type": "integer" } },
          "X-RateLimit-Reset": { "description": "Seconds until the bucket is full again.", "schema": { "type": "integer" } },
          "X-RateLimit-Bucket": { "description": "Which tier is binding: tenant, credential or expensive.", "schema": { "type": "string" } }
        },
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": ["code", "message"],
        "properties": {
          "code": { "type": "string", "description": "Stable machine-readable identifier. Branch on this, not on the message." },
          "message": { "type": "string", "description": "Human-readable explanation. 5xx messages are deliberately generic: an upstream error string leaks topology." },
          "request_id": { "type": "string" }
        }
      },
...
      "ReasonRequest": {
        "type": "object",
        "properties": {
          "reason": { "type": "string", "description": "Recorded in the audit log. Supply one: the audit trail is read by people who were not there." },
          "initiated_by": { "type": "string" }
        }
      },
...
    }
  }
}